functions/blog_post_image_generator/requirements.txt
boto3==1.28.42
PyGithub==1.59.1
toml==0.10.2